Add blacklisted associations on Ember Blueprints
In case someone wants to blacklist an association to be returned in the
JSON for Ember-Data, you can specify it on the configuration.
Example:

User -> Auth
User -> Jwt

By NO MEANS the API should return the Auth data and Jwt tokens. This can
now be blacklisted and will never show up in the JSON.