Remove heavy collection on Plugin model
By removing the `installations` collection on the Plugins's `toJSON`
method, we ensure that neither the server or the client will blow up if
there's a plugin with thousands of intallations.
The PluginInstallations model will be used only for statistics. If the
client need to know how many times a plugin has been installed, there's
the `installationCount` property on this model.